feat: implement multi-step area creation with action and reaction selection
